Attempting to create long-lived connection.
The goal is to make the client respond quicker to updates and to not
overload the server/network with requests.  There are somd fundamental
problems with the current attempt here.

Currently, the first connection doesn't open right away.

I'm not sure of the side effecs of not closing the on-timeout handler
that should respond without update.

When doing a submit, the immediate update event is lost (doesn't trigger
a response), but that might be reasonable.  It does show that the
timeout response contains the updated data.